requirements.txt
dj-database-url==0.4.2
Django==1.10.5
django-fontawesome==0.3.1
djangorestframework==3.5.4
prices==0.5.9
psycopg2==2.6.2
python-http-client==2.2.1
PyYAML==3.12
satchless==1.1.3
sendgrid==0.3.1
smtpapi==0.3.1
source==1.2.0
virtualenv==15.1.0
whitenoise==3.3.0
